+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 
+- Show the project name instead of the filter when there is only one project.
+
 ## [1.0.0] - 2026-04-09
 
 - Initial release.